third_party/blink/web_tests/external/wpt/css/css-text-decor/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-emphasis-position has initial value auto
  assert_equals: expected "auto" but got "over"
Harness: the test ran to completion.
